refactor: flatten load-env module
