+++ src/broadcast-reminder.ts
@@ -4,7 +4,7 @@ import { getDiscordClient } from './clients';
 import { formatReminderMessage, getReminderByTime, removeReminders } from './commands/reminder/reminder-utils';
 import { getReminderChannel } from './commands/serverSettings/server-utils';
 import { getCurrentUnixTime } from './utils/date-utils';
-import { loadEnv } from './utils/loadEnv';
+import { loadEnv } from './utils/load-env';
 import { logger } from './utils/logger';
 
 const broadcastReminder = async () => {
