+++ src/clients/prisma.ts
@@ -2,7 +2,7 @@ import { Pool, neonConfig } from '@neondatabase/serverless';
 import { PrismaNeon } from '@prisma/adapter-neon';
 import { PrismaClient } from '@prisma/client';
 import ws from 'ws';
-import { loadEnv } from '../utils/loadEnv';
+import { loadEnv } from '../utils/load-env';
 
 let prisma: PrismaClient;
