+++ src/scripts/delete-guild-commands.ts
@@ -1,7 +1,7 @@
 import { Result } from 'oxide.ts';
 import { deployGuildCommands } from '../commands/deploy-command';
 import { getCurrentUnixTime } from '../utils/date-utils';
-import { loadEnv } from '../utils/loadEnv';
+import { loadEnv } from '../utils/load-env';
 import { logger } from '../utils/logger';
 
 const deploy = async () => {
